Moved additions the Melvor types to their own files
