+++ src/modifiers/CustomModifiersManager.ts
@@ -5,281 +5,7 @@ import { MonsterTypeMappingManager } from './MonsterTypeMappingManager';
 import { CustomModifiersCalculator } from './CustomModifiersCalculator'
 
 /** For typescript intellisense and not throwing errors */
-declare global {
-    interface StandardModifierObject<Standard> {
-        increasedMaxHitPercentToCombatAreaMonsters: Standard,
-        decreasedMaxHitPercentToCombatAreaMonsters: Standard,
-        increasedMaxHitFlatToCombatAreaMonsters: Standard,
-        decreasedMaxHitFlatToCombatAreaMonsters: Standard,
-        increasedMinHitBasedOnMaxHitToCombatAreaMonsters: Standard,
-        decreasedMinHitBasedOnMaxHitToCombatAreaMonsters: Standard,
-        increasedFlatMinHitToCombatAreaMonsters: Standard,
-        decreasedFlatMinHitToCombatAreaMonsters: Standard,
-
-        increasedMaxHitPercentToSlayerAreaMonsters: Standard,
-        decreasedMaxHitPercentToSlayerAreaMonsters: Standard,
-        increasedMaxHitFlatToSlayerAreaMonsters: Standard,
-        decreasedMaxHitFlatToSlayerAreaMonsters: Standard,
-        increasedMinHitBasedOnMaxHitToSlayerAreaMonsters: Standard,
-        decreasedMinHitBasedOnMaxHitToSlayerAreaMonsters: Standard,
-        increasedFlatMinHitToSlayerAreaMonsters: Standard,
-        decreasedFlatMinHitToSlayerAreaMonsters: Standard,
-
-        increasedMaxHitPercentToDungeonMonsters: Standard,
-        decreasedMaxHitPercentToDungeonMonsters: Standard,
-        increasedMaxHitFlatToDungeonMonsters: Standard,
-        decreasedMaxHitFlatToDungeonMonsters: Standard,
-        increasedMinHitBasedOnMaxHitToDungeonMonsters: Standard,
-        decreasedMinHitBasedOnMaxHitToDungeonMonsters: Standard,
-        increasedFlatMinHitToDungeonMonsters: Standard,
-        decreasedFlatMinHitToDungeonMonsters: Standard,
-
-        increasedMaxHitPercentToSlayerTasks: Standard,
-        decreasedMaxHitPercentToSlayerTasks: Standard,
-        increasedMaxHitFlatToSlayerTasks: Standard,
-        decreasedMaxHitFlatToSlayerTasks: Standard,
-        increasedMinHitBasedOnMaxHitToSlayerTasks: Standard,
-        decreasedMinHitBasedOnMaxHitToSlayerTasks: Standard,
-        increasedFlatMinHitToSlayerTasks: Standard,
-        decreasedFlatMinHitToSlayerTasks: Standard,
-
-        increasedMaxHitPercentAgainstBosses: Standard,
-        decreasedMaxHitPercentAgainstBosses: Standard,
-        increasedMaxHitFlatAgainstBosses: Standard,
-        decreasedMaxHitFlatAgainstBosses: Standard,
-        increasedMinHitBasedOnMaxHitAgainstBosses: Standard,
-        decreasedMinHitBasedOnMaxHitAgainstBosses: Standard,
-        increasedFlatMinHitAgainstBosses: Standard,
-        decreasedFlatMinHitAgainstBosses: Standard
-
-        /** Percentage */
-        increasedGlobalSkillXPPerLevel: Standard,
-        /** Percentage */
-        decreasedGlobalSkillXPPerLevel: Standard,
-        /** Increases total xp gain for all skills (so unaffected by xp multipliers!) */
-        increasedFlatGlobalSkillXP: Standard,
-        /** Decreases total xp gain for all skills (so unaffected by xp multipliers!) */
-        decreasedFlatGlobalSkillXP: Standard,
-        /** Increases total xp gain for all skills (so unaffected by xp multipliers!) */
-        increasedFlatGlobalSkillXPPerSkillLevel: Standard,
-        /** Decreases total xp gain for all skills (so unaffected by xp multipliers!) */
-        decreasedFlatGlobalSkillXPPerSkillLevel: Standard
-    }
-
-    interface SkillModifierObject<Skill> {
-        /** Percentage */
-        increasedSkillXPPerSkillLevel: Skill,
-        /** Percentage */
-        decreasedSkillXPPerSkillLevel: Skill,
-        /** Increases total xp gain for given skill (so unaffected by xp multipliers!) */
-        increasedFlatSkillXP: Skill,
-        /** Decreased total xp gain for given skill (so unaffected by xp multipliers!) */
-        decreasedFlatSkillXP: Skill,
-        /** Increases total xp gain for given skill (so unaffected by xp multipliers!) */
-        increasedFlatSkillXPPerSkillLevel: Skill,
-        /** Decreased total xp gain for given skill (so unaffected by xp multipliers!) */
-        decreasedFlatSkillXPPerSkillLevel: Skill,
-    }
-
-    interface PlayerModifiers {
-        increasedMaxHitPercentToCombatAreaMonsters: number,
-        decreasedMaxHitPercentToCombatAreaMonsters: number,
-        increasedMaxHitFlatToCombatAreaMonsters: number,
-        decreasedMaxHitFlatToCombatAreaMonsters: number,
-        increasedMinHitBasedOnMaxHitToCombatAreaMonsters: number,
-        decreasedMinHitBasedOnMaxHitToCombatAreaMonsters: number,
-        increasedFlatMinHitToCombatAreaMonsters: number,
-        decreasedFlatMinHitToCombatAreaMonsters: number,
-
-        increasedMaxHitPercentToSlayerAreaMonsters: number,
-        decreasedMaxHitPercentToSlayerAreaMonsters: number,
-        increasedMaxHitFlatToSlayerAreaMonsters: number,
-        decreasedMaxHitFlatToSlayerAreaMonsters: number,
-        increasedMinHitBasedOnMaxHitToSlayerAreaMonsters: number,
-        decreasedMinHitBasedOnMaxHitToSlayerAreaMonsters: number,
-        increasedFlatMinHitToSlayerAreaMonsters: number,
-        decreasedFlatMinHitToSlayerAreaMonsters: number,
-
-        increasedMaxHitPercentToDungeonMonsters: number,
-        decreasedMaxHitPercentToDungeonMonsters: number,
-        increasedMaxHitFlatToDungeonMonsters: number,
-        decreasedMaxHitFlatToDungeonMonsters: number,
-        increasedMinHitBasedOnMaxHitToDungeonMonsters: number,
-        decreasedMinHitBasedOnMaxHitToDungeonMonsters: number,
-        increasedFlatMinHitToDungeonMonsters: number,
-        decreasedFlatMinHitToDungeonMonsters: number,
-
-        increasedMaxHitPercentToSlayerTasks: number,
-        decreasedMaxHitPercentToSlayerTasks: number,
-        increasedMaxHitFlatToSlayerTasks: number,
-        decreasedMaxHitFlatToSlayerTasks: number,
-        increasedMinHitBasedOnMaxHitToSlayerTasks: number,
-        decreasedMinHitBasedOnMaxHitToSlayerTasks: number,
-        increasedFlatMinHitToSlayerTasks: number,
-        decreasedFlatMinHitToSlayerTasks: number,
-
-        increasedMaxHitPercentAgainstBosses: number,
-        decreasedMaxHitPercentAgainstBosses: number,
-        increasedMaxHitFlatAgainstBosses: number,
-        decreasedMaxHitFlatAgainstBosses: number,
-        increasedMinHitBasedOnMaxHitAgainstBosses: number,
-        decreasedMinHitBasedOnMaxHitAgainstBosses: number,
-        increasedFlatMinHitAgainstBosses: number,
-        decreasedFlatMinHitAgainstBosses: number
-
-        /** Percentage */
-        increasedGlobalSkillXPPerLevel: number,
-        /** Percentage */
-        decreasedGlobalSkillXPPerLevel: number,
-        /** Increases total xp gain for all skills (so unaffected by xp multipliers!) */
-        increasedFlatGlobalSkillXP: number,
-        /** Decreases total xp gain for all skills (so unaffected by xp multipliers!) */
-        decreasedFlatGlobalSkillXP: number,
-        /** Increases total xp gain for all skills (so unaffected by xp multipliers!) */
-        increasedFlatGlobalSkillXPPerSkillLevel: number,
-        /** Decreases total xp gain for all skills (so unaffected by xp multipliers!) */
-        decreasedFlatGlobalSkillXPPerSkillLevel: number
-    }
-
-    interface CombatModifierObject<Standard> {
-        increasedChanceToApplySlowOnSpawn: Standard,
-        decreasedChanceToApplySlowOnSpawn: Standard,
-        increasedChanceToApplyStunOnSpawn: Standard,
-        decreasedChanceToApplyStunOnSpawn: Standard,
-        increasedChanceToApplyPoisonOnSpawn: Standard,
-        decreasedChanceToApplyPoisonOnSpawn: Standard,
-        increasedChanceToApplyDeadlyPoisonOnSpawn: Standard,
-        decreasedChanceToApplyDeadlyPoisonOnSpawn: Standard,
-
-        /** Flag (by the custom effect) that effect has been applied, it's NOT an increasing value */
-        deathMark: Standard,
-        increasedDeathMarkOnHit: Standard,
-        increasedChanceToApplyStackOfDeathMark: Standard,
-        decreasedChanceToApplyStackOfDeathMark: Standard,
-        increasedDeathMarkImmunity: Standard,
-        decreasedDeathMarkImmunity: Standard,
-
-        increasedDamageTakenFromAirSpells: Standard,
-        decreasedDamageTakenFromAirSpells: Standard,
-        increasedDamageTakenFromWaterSpells: Standard,
-        decreasedDamageTakenFromWaterSpells: Standard,
-        increasedDamageTakenFromEarthSpells: Standard,
-        decreasedDamageTakenFromEarthSpells: Standard,
-        increasedDamageTakenFromFireSpells: Standard,
-        decreasedDamageTakenFromFireSpells: Standard,
-
-        humanTraitApplied: Standard,
-        dragonTraitApplied: Standard,
-        undeadTraitApplied: Standard,
-
-        increasedDamageAgainstHumans: Standard,
-        decreasedDamageAgainstHumans: Standard,
-        increasedMaxHitPercentAgainstHumans: Standard,
-        decreasedMaxHitPercentAgainstHumans: Standard,
-        increasedMaxHitFlatAgainstHumans: Standard,
-        decreasedMaxHitFlatAgainstHumans: Standard,
-        increasedMinHitBasedOnMaxHitAgainstHumans: Standard,
-        decreasedMinHitBasedOnMaxHitAgainstHumans: Standard,
-        increasedFlatMinHitAgainstHumans: Standard,
-        decreasedFlatMinHitAgainstHumans: Standard,
-
-        increasedDamageAgainstDragons: Standard,
-        decreasedDamageAgainstDragons: Standard,
-        increasedMaxHitPercentAgainstDragons: Standard,
-        decreasedMaxHitPercentAgainstDragons: Standard,
-        increasedMaxHitFlatAgainstDragons: Standard,
-        decreasedMaxHitFlatAgainstDragons: Standard,
-        increasedMinHitBasedOnMaxHitAgainstDragons: Standard,
-        decreasedMinHitBasedOnMaxHitAgainstDragons: Standard,
-        increasedFlatMinHitAgainstDragons: Standard,
-        decreasedFlatMinHitAgainstDragons: Standard,
-
-        increasedDamageAgainstUndead: Standard,
-        decreasedDamageAgainstUndead: Standard,
-        increasedMaxHitPercentAgainstUndead: Standard,
-        decreasedMaxHitPercentAgainstUndead: Standard,
-        increasedMaxHitFlatAgainstUndead: Standard,
-        decreasedMaxHitFlatAgainstUndead: Standard,
-        increasedMinHitBasedOnMaxHitAgainstUndead: Standard,
-        decreasedMinHitBasedOnMaxHitAgainstUndead: Standard,
-        increasedFlatMinHitAgainstUndead: Standard,
-        decreasedFlatMinHitAgainstUndead: Standard
-    }
-
-    interface CombatModifiers {
-        increasedChanceToApplySlowOnSpawn: number,
-        decreasedChanceToApplySlowOnSpawn: number,
-        increasedChanceToApplyStunOnSpawn: number,
-        decreasedChanceToApplyStunOnSpawn: number,
-        increasedChanceToApplyPoisonOnSpawn: number,
-        decreasedChanceToApplyPoisonOnSpawn: number,
-        increasedChanceToApplyDeadlyPoisonOnSpawn: number,
-        decreasedChanceToApplyDeadlyPoisonOnSpawn: number,
-
-        deathMark: number,
-        increasedDeathMarkOnHit: number,
-        increasedChanceToApplyStackOfDeathMark: number,
-        decreasedChanceToApplyStackOfDeathMark: number,
-        increasedDeathMarkImmunity: number,
-        decreasedDeathMarkImmunity: number,
-
-        increasedDamageTakenFromAirSpells: number,
-        decreasedDamageTakenFromAirSpells: number,
-        increasedDamageTakenFromWaterSpells: number,
-        decreasedDamageTakenFromWaterSpells: number,
-        increasedDamageTakenFromEarthSpells: number,
-        decreasedDamageTakenFromEarthSpells: number,
-        increasedDamageTakenFromFireSpells: number,
-        decreasedDamageTakenFromFireSpells: number,
-
-        humanTraitApplied: number,
-        dragonTraitApplied: number,
-        undeadTraitApplied: number,
-
-        increasedDamageAgainstHumans: number,
-        decreasedDamageAgainstHumans: number,
-        increasedMaxHitPercentAgainstHumans: number,
-        decreasedMaxHitPercentAgainstHumans: number,
-        increasedMaxHitFlatAgainstHumans: number,
-        decreasedMaxHitFlatAgainstHumans: number,
-        increasedMinHitBasedOnMaxHitAgainstHumans: number,
-        decreasedMinHitBasedOnMaxHitAgainstHumans: number,
-        increasedFlatMinHitAgainstHumans: number,
-        decreasedFlatMinHitAgainstHumans: number,
-
-        increasedDamageAgainstDragons: number,
-        decreasedDamageAgainstDragons: number,
-        increasedMaxHitPercentAgainstDragons: number,
-        decreasedMaxHitPercentAgainstDragons: number,
-        increasedMaxHitFlatAgainstDragons: number,
-        decreasedMaxHitFlatAgainstDragons: number,
-        increasedMinHitBasedOnMaxHitAgainstDragons: number,
-        decreasedMinHitBasedOnMaxHitAgainstDragons: number,
-        increasedFlatMinHitAgainstDragons: number,
-        decreasedFlatMinHitAgainstDragons: number,
-
-        increasedDamageAgainstUndead: number,
-        decreasedDamageAgainstUndead: number,
-        increasedMaxHitPercentAgainstUndead: number,
-        decreasedMaxHitPercentAgainstUndead: number,
-        increasedMaxHitFlatAgainstUndead: number,
-        decreasedMaxHitFlatAgainstUndead: number,
-        increasedMinHitBasedOnMaxHitAgainstUndead: number,
-        decreasedMinHitBasedOnMaxHitAgainstUndead: number,
-        increasedFlatMinHitAgainstUndead: number,
-        decreasedFlatMinHitAgainstUndead: number
-    }
-
-    interface Character {
-        isHuman: boolean,
-        isDragon: boolean,
-        isUndead: boolean
-    }
-
-    interface Game {
-        deathMarkEffect: StackingEffect
-    }
-}
+require('./definitions')
 
 /**
  * Patches different sections of the code, in order to integrate custom modifiers
