+++ src/modifiers/definitions/Character.d.ts
@@ -1,0 +1,9 @@
+declare global {
+    interface Character {
+        isHuman: boolean,
+        isDragon: boolean,
+        isUndead: boolean
+    }
+}
+
+export { };
