+++ src/modifiers/definitions/Game.d.ts
@@ -1,0 +1,7 @@
+declare global {
+    interface Game {
+        deathMarkEffect: StackingEffect
+    }
+}
+
+export { };
